Replace all old tests with minimal new ones
